refactor: each commands in separate folder
