+++ cmd/cat/main.ts
@@ -1,0 +1,5 @@
+import { fetch, Options } from "rclone";
+
+export function cat(location: string, flags?: Options): Promise<Response> {
+  return fetch(location, flags);
+}
