+++ cmd/cat/main_test.ts
@@ -1,0 +1,6 @@
+import { assert } from "../../dev_deps.ts";
+import { cat } from "./main.ts";
+
+Deno.test("cat", () => {
+  assert(cat);
+});
