+++ cmd/copy/main_test.ts
@@ -1,0 +1,6 @@
+import { assert } from "../../dev_deps.ts";
+import { copy } from "./main.ts";
+
+Deno.test("copy", () => {
+  assert(copy);
+});
