+++ cmd/ls/main_test.ts
@@ -1,0 +1,6 @@
+import { assert } from "../../dev_deps.ts";
+import { ls } from "./main.ts";
+
+Deno.test("ls", () => {
+  assert(ls);
+});
