+++ cmd/lsd/main_test.ts
@@ -1,0 +1,6 @@
+import { assert } from "../../dev_deps.ts";
+import { lsd } from "./main.ts";
+
+Deno.test("lsd", () => {
+  assert(lsd);
+});
