+++ cmd/lsf/main_test.ts
@@ -1,0 +1,6 @@
+import { assert } from "../../dev_deps.ts";
+import { lsf } from "./main.ts";
+
+Deno.test("lsf", () => {
+  assert(lsf);
+});
