+++ cmd/lsjson/main_test.ts
@@ -1,0 +1,6 @@
+import { assert } from "../../dev_deps.ts";
+import { lsjson } from "./main.ts";
+
+Deno.test("lsjson", () => {
+  assert(lsjson);
+});
