+++ cmd/lsl/main_test.ts
@@ -1,0 +1,6 @@
+import { assert } from "../../dev_deps.ts";
+import { lsl } from "./main.ts";
+
+Deno.test("lsl", () => {
+  assert(lsl);
+});
