+++ cmd/rcat/main_test.ts
@@ -1,0 +1,6 @@
+import { assert } from "../../dev_deps.ts";
+import { rcat } from "./main.ts";
+
+Deno.test("rcat", () => {
+  assert(rcat);
+});
